Add UInt addition
This involves a significant architectural change:
the Buffer no longer subclasses Array,
as many operations were made harder because of it.

In particular, you could not reset the list of bits
to a new array.
You instead had to rely on awkward splice() combinations.

It got worse, as splice() created new objects
with the same constructor as the current class.
It made using splice impossible in the constructor,
as it would loop indefinitely until the call stack overflew.

Composition instead of inheritance makes all these issues go away.