Add De Morgan 2 and OR associativity with NOT
We initially assumed that De Morgan was not valuable to implement,
because it increased the number of binary operations
in the boolean expression.

However, the failure to reduce XOR associativity
with respect to NOT reduction that was discovered
in 6b457c8dea67b256a6bd711800e3312af4004581
has a symmetrical issue in failing to reduce
OR associativity in a De Morgan situation:
the expression (¬(a∧b)∨¬a) failed to reduce
instead of  reducing to (¬a∨¬b).

Therefore, we now always convert a De Morgan 2 situation
into a disjunction that can then be reduced by other rules.