Add De Morgan 1 and AND associativity with NOT
This change is the symmetrical equivalent of the change in
eb0a7dd9a1aa9161d130117f68e042f37138a10a,
but with AND instead of OR:

The expression (¬(a∨b)∧¬a) failed to reduce
instead of reducing to (¬a∧¬b).

Therefore, we now always convert a De Morgan 1 situation
into a conjunction that can then be reduced by other rules.