Implement associative bit probability computation
This implementation focuses on:

- Accurate probability computation (no approximation),
- Boolean expression that are not independent.

The much easier computation is that of independent expressions.
Traditionally, accuracy is sacrificed for dependent ones,
by the use of Monte-Carlo methods.

After all, this is equivalent to SAT,
which is famously the first NP-complete problem found.
Therefore, exact results quickly grow exponentially hard to compute
as the size of the problem grows linearly.
(That is the Exponential Time Hypothesis.)

Starting with this ensures that we can demonstrate
accurate results for cases with little dependencies between expressions.
To avoid infinite loops, we will later handle cases too large to compute
with said Monte-Carlo methods.

On the plus side, we will only get approximate results
as the size of dependencies grow,
instead of the size of variables.