Test independent boolean expressions’ operation probability
I assumed yesterday that I would still require work
to get independent expressions working,
because I thought the loop was never executed.

But 2^0 = 1, so the loop is actually executed once,
such that there is no change to the algorithm:
independent expressions work with no modification.

So I simply added tests for them.